requirements.txt
django==4.2.23
psycopg2-binary==2.9.10
djangorestframework==3.15.2
gunicorn==21.2.0